move contributor verificaiton to own folder
