+++ src/pages/Admin/Charity/OtherSettings/DonorVerification/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./OtherSettings";
